mcp-server/pyproject.toml
[project]
name = "meal-planning-mcp-server"
version = "0.1.0"
description = "MCP server for meal planning with recipe finder and shopping list tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.1.0",
    "pydantic>=2.0.0",
    "uvloop>=0.19.0",
    "faiss-cpu>=1.8.0; platform_machine != 'arm64' or platform_system != 'Darwin'",
    "faiss-cpu>=1.8.0.post1; platform_machine == 'arm64' and platform_system == 'Darwin'",
    "numpy>=1.26.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["meal_planning_mcp*"]